EVM-1167: split storage_types into focused modules
