+++ zk_ee/src/storage_types/constants.rs
@@ -1,0 +1,2 @@
+/// Maximum number of topics allowed per event.
+pub const MAX_EVENT_TOPICS: usize = 4;
